.github/actions/issue-analyzer/requirements.txt
anthropic>=0.18.1
PyGithub>=2.2.0